+++ packages/date-input/src/Component.screenshots.test.tsx
@@ -1,0 +1,96 @@
+import { setupScreenshotTesting, createSpriteStorybookUrl } from '../../screenshot-utils';
+
+const screenshotTesting = setupScreenshotTesting({
+    it,
+    beforeAll,
+    afterAll,
+    expect,
+});
+
+describe(
+    'DateInput | sizes',
+    screenshotTesting({
+        cases: [
+            [
+                'sprite',
+                createSpriteStorybookUrl({
+                    componentName: 'DateInput',
+                    size: { width: 350, height: 400 },
+                    knobs: {
+                        value: ['30.12.2005', '01.06.2022'],
+                        size: ['s', 'm', 'l', 'xl'],
+                    },
+                }),
+            ],
+        ],
+        screenshotOpts: {
+            fullPage: true,
+        },
+    }),
+);
+
+describe(
+    'DateInput | error',
+    screenshotTesting({
+        cases: [
+            [
+                'sprite',
+                createSpriteStorybookUrl({
+                    componentName: 'DateInput',
+                    size: { width: 350, height: 400 },
+                    knobs: {
+                        value: ['30.12.2005', '01.06.2022'],
+                        error: [false, true],
+                    },
+                }),
+            ],
+        ],
+        screenshotOpts: {
+            fullPage: true,
+        },
+    }),
+);
+
+describe(
+    'DateInput | hint',
+    screenshotTesting({
+        cases: [
+            [
+                'sprite',
+                createSpriteStorybookUrl({
+                    componentName: 'DateInput',
+                    size: { width: 350, height: 400 },
+                    knobs: {
+                        value: ['30.12.2005', '01.06.2022'],
+                        hint: [null, 'Hint'],
+                    },
+                }),
+            ],
+        ],
+        screenshotOpts: {
+            fullPage: true,
+        },
+    }),
+);
+
+describe(
+    'DateInput | label',
+    screenshotTesting({
+        cases: [
+            [
+                'sprite',
+                createSpriteStorybookUrl({
+                    componentName: 'DateInput',
+                    size: { width: 350, height: 400 },
+                    knobs: {
+                        value: ['30.12.2005', '01.06.2022'],
+                        label: [null, 'Label'],
+                    },
+                }),
+            ],
+        ],
+        screenshotOpts: {
+            fullPage: true,
+        },
+    }),
+);