+++ packages/tooltip/src/component.responsive.tsx
@@ -114,8 +114,10 @@ export const TooltipResponsive: FC<TooltipResponsiveProps> = ({
     ) : (
         <Tooltip
             {...restProps}
+            open={open}
             content={content}
             onOpen={handleOpen}
+            onClose={handleClose}
             targetClassName={targetClassName}
             targetRef={targetRef}
         >