+++ packages/typography/src/index.ts
@@ -2,3 +2,4 @@ export * from './component';
 export * from './presets';
 export { TitleProps } from './title';
 export { TextProps } from './text';
+export { Color } from './colors';