+++ packages/password-input/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.9](https://github.com/core-ds/core-components/compare/@alfalab/core-components-password-input@2.2.8...@alfalab/core-components-password-input@2.2.9) (2022-06-24)
+
+
+### Bug Fixes
+
+* **icon-button:** fix disabled colors ([#104](https://github.com/core-ds/core-components/issues/104)) ([3f03849](https://github.com/core-ds/core-components/commit/3f038495bb63f72cd81ceeedbe55b52119581d57))
+
+
+
+
+
+## [2.2.8](https://github.com/core-ds/core-components/compare/@alfalab/core-components-password-input@2.2.7...@alfalab/core-components-password-input@2.2.8) (2022-06-23)
+
+**Note:** Version bump only for package @alfalab/core-components-password-input
+
+
+
+
+
+## [2.2.7](https://github.com/core-ds/core-components/compare/@alfalab/core-components-password-input@2.2.6...@alfalab/core-components-password-input@2.2.7) (2022-06-20)
+
+
+### Bug Fixes
+
+* **icon-button:** fix loading ([#105](https://github.com/core-ds/core-components/issues/105)) ([0b133f0](https://github.com/core-ds/core-components/commit/0b133f042e86702ec2861915f2cdbcdbad9ca905))
+
+
+
+
+
 ## [2.2.6](https://github.com/core-ds/core-components/compare/@alfalab/core-components-password-input@2.2.5...@alfalab/core-components-password-input@2.2.6) (2022-06-03)
 
 **Note:** Version bump only for package @alfalab/core-components-password-input