+++ packages/themes/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.2.2](https://github.com/core-ds/core-components/compare/@alfalab/core-components-themes@8.2.1...@alfalab/core-components-themes@8.2.2) (2022-06-24)
+
+
+### Bug Fixes
+
+* **icon-button:** fix disabled colors ([#104](https://github.com/core-ds/core-components/issues/104)) ([3f03849](https://github.com/core-ds/core-components/commit/3f038495bb63f72cd81ceeedbe55b52119581d57))
+
+
+
+
+
+## [8.2.1](https://github.com/core-ds/core-components/compare/@alfalab/core-components-themes@8.2.0...@alfalab/core-components-themes@8.2.1) (2022-06-23)
+
+
+### Bug Fixes
+
+* **button:** fix color secondary ([#92](https://github.com/core-ds/core-components/issues/92)) ([57b8e66](https://github.com/core-ds/core-components/commit/57b8e661b8f54acdfd0c235f58ebd59d66c116a4))
+
+
+
+
+
 # [8.2.0](https://github.com/core-ds/core-components/compare/@alfalab/core-components-themes@8.1.5...@alfalab/core-components-themes@8.2.0) (2022-06-08)